DSPDC-725 Initial Commit (#5)
* DSPDC-725 Initial Commit
Adding a new object to the EncodeEntity collection to add FunctionalCharacterizationExperiment. This dataset was previously assumed to be included under the "type=Experiment" query.
Also added code to the ExtractionPipeline script to store a new parent entity FunctionalCharacterizationExperiment and also added code to ensure we're pulling file data from that set of experiments.

* QH DSPDC-725 Updates
As written, the two calls to extractLinkedEntities at the end of this script would probably write their outputs to the same prefix which would cause issues.
The extractLinkedEntities function expects a "SCollection[JsonObject]" for the linkingEntities parameter so we can use a unionAll to combine the experiments and fcExperiments entities and extract from both of them.

* QH DSPDC-725 Bugfix
SCollection.unionAll expects an iterable list

* QH DSPDC-725 - Bugfix
Incorrectly assumed you can just feed in a list of lists or wrap them in a brackets to turn them into an iterable list (not in Scala!).
Scala requires you build collections via apply on their companion objects, so I updated the script to wrap the two items with List().

* QH DSPDC-725 Optimization
Optimized the code to ensure we are separating batches within the replicates stream into normal experiments and functional-characterization-experiments.
Then feeding each of those partitions into "extractLinkedEntities".

* QH DSPDC-725 - Bugfix + Unit Test
Added framework for adding unit tests (build.sbt)
Removed the "manyReferences" parameter since its original use case was to address the different types of experiments which is now solved with this ticket (EncodeExtractions)
Added a function to ExtractionPipeline to determine whether an individual experiment within a replicate stream is a "functional-characterization-experiment" or just a normal "experiment" by returning a boolean value. Passing this new function into a partition call that acts on the replicates stream to parse expeiments.
Added a unit test with two cases (fcExperiment vs experiment) to test the functionality of the new method.